New improved GenericMessage init_subclass
Now we define the message_type, if any, on the class definition attributes.
For example:

```
class Hello(GenericMessage, message_type='OFPT_HELLO'):
    pass
```

So, you can just use a string to define the message type and also you do not
have to worry about importing the Type Enum, setting it at the 'right place'
and also you won't forget to set the message_type (if you do so, the class
creation will raise an error).